hooks: better handle config without gitdir
When the post-command hook runs, we could be in several custom states:

1. The command is a regular builtin, but the repo setup is incomplete.
   (Example: "git version", "git rev-parse HEAD")

2. The command is a dashed command, but the top level process uses a space
   in its call. (Example: "git http-request")

3. The command is an alias that goes to a builtin.

4. The command has no arguments or is only helptext.

Each of these cases leads to a place where we previously had not loaded the
postCommand.strategy config and would execute the post-command hook without
looking for a sentinel file.

There are two fixes here:

First, we use early config parsing so we can get config details without
fully setting up the repository. This is how core.hooksPath works in these
situations.

Second, we need to make sure handle_hook_replacement() is called even when
the repo's gitdir is NULL. This requires a small amount of care to say that
a sentinel file cannot exist if the gitdir isn't set, as we would not have
written one without initializing the gitdir.

This gets all of the desired behaviors we want for this setting without
doing anything extreme with how pre- and post-command hooks run otherwise.

+++ hook.c
@@ -173,8 +173,14 @@ static int write_post_index_change_sentinel(struct repository *r)
  */
 static int post_index_change_sentinel_exists(struct repository *r)
 {
-	char *path = get_post_index_change_sentinel_name(r);
+	char *path;
 	int res = 1;
+
+	/* It can't exist if we don't have a gitdir. */
+	if (!r->gitdir)
+		return 0;
+
+	path = get_post_index_change_sentinel_name(r);
 
 	if (unlink(path)) {
 		if (is_missing_file_error(errno))
@@ -187,6 +193,21 @@ static int post_index_change_sentinel_exists(struct repository *r)
 	return res;
 }
 
+static int check_worktree_change(const char *key, const char *value,
+				       UNUSED const struct config_context *ctx,
+				       void *data)
+{
+	int *enabled = data;
+
+	if (!strcmp(key, "postcommand.strategy") &&
+	    !strcasecmp(value, "worktree-change")) {
+		*enabled = 1;
+		return 1;
+	}
+
+	return 0;
+}
+
 /**
  * See if we can replace the requested hook with an internal behavior.
  * Returns 0 if the real hook should run. Returns nonzero if we instead
@@ -196,9 +217,11 @@ static int handle_hook_replacement(struct repository *r,
 				   const char *hook_name,
 				   struct strvec *args)
 {
-	const char *strval;
-	if (repo_config_get_string_tmp(r, "postcommand.strategy", &strval) ||
-	    strcasecmp(strval, "worktree-change"))
+	int enabled = 0;
+
+	read_early_config(r, check_worktree_change, &enabled);
+
+	if (!enabled)
 		return 0;
 
 	if (!strcmp(hook_name, "post-index-change")) {
@@ -227,7 +250,7 @@ static void list_hooks_add_default(struct repository *r, const char *hookname,
 	struct hook *h;
 
 	/* Interject hook behavior depending on strategy. */
-	if (r && r->gitdir && options &&
+	if (r && options &&
 	    handle_hook_replacement(r, hookname, &options->args))
 		return;
